+++ NEWS.md
@@ -9,6 +9,8 @@ population previously gave an error.
 specified were giving an error (#44). `aux_regression` and `aux_by` have no 
 effect for exponential models since there are no auxiliary (shape) parameters 
 and are ignored, now with a warning.
+* Fix: Avoid an error when trying to fit M-spline models combining IPD and AgD 
+in R versions prior to 4.1.0, due to integer coercion of factors by `c()`.
 
 # multinma 0.7.1